Math - Sin Wave
This update adds the first of the Math library, the sin wave. The sin wave creates an object that tracks positioning in a sin wave. This includes the sin wave amplitude and frequency. This class also has the ability to follow a sin wave along an angled line (e.g., sin wave that is angled at 30 degrees).

Additionally, the particle class now takes a sin wave as a parameter. There will eventually be an update to expand this to allow different kinds of patterns (e.g., circling, spiraling, etc.).

The main file now is updated to show examples of particles using the sin wave at 0 and 180 degrees.

+++ Bebop/Source/BebopCore/Objects/CircleObject.cpp
@@ -32,7 +32,7 @@ namespace Bebop { namespace Objects
    //    N/A
    //
    //************************************************************************************************************
-   CircleObject::CircleObject(const int aCoordinateX, const int aCoordinateY, const int aRadius) :
+   CircleObject::CircleObject(const float aCoordinateX, const float aCoordinateY, const int aRadius) :
       Object(aCoordinateX, aCoordinateY, aRadius, aRadius), mRadius(aRadius)
    {
    }
